backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
pandas
openpyxl
flask-jwt-extended
